Add gentoolkitNonZeroExit exception
This exception should only be used when an error is not fatal and the
absence of information means no data was found.  The return_code
parameter can be used to set the exit value.

Have the equery list module when in quiet mode return an exit status of
3.  This is needed to prevent 'euse -I' from printing many 'No installed
packages matching ...' error messages.