Apply latest changes from Jared Hancock
Fixes described issues from:
http://bugs.gentoo.org/show_bug.cgi?id=259318#c88

"It's now more vocal when handling global flags, and does not indicate
make.conf was edited if it was not. Actually, more specifically, it
does not modify make.conf unnecessarily. (As before, it modified
make.conf every time regardless of whether or not it changed anything)

Also fixed the issue with removing and disabling multiple flags for a
package."