Fix extend_realpaths in FileOwner class. (Bug 96515)
Removed the os.path.islink check since it oly returns true if the last
part of the path is a symbolic link.  We want to add the realpath to the
list if it already doesn't exist in the list, since that indicates that
something in the path is a symbolic link.