revdep_rebuild/analyse.py: Fix the bug that caused some breakage to not be detected.
scan_files() was not tracking the filenames and needed correctly.
This was causing data loss for files that need to be checked.
I found there could be several filenames for any given soname.
Each filename has a set of needed libs to track.
find_broken2 and main_checks2 needed updating for the different scanned_files structure.