gentoolkit: Fix to allow package names of all digits. Bug 562952
According to PMS, "A package name may contain any of the characters
[A-Za-z0-9+_-]. It must not begin with a hyphen or a plus sign, and must
not end in a hyphen followed by anything matching the version syntax".

This specification means that a package name can be composed of all
digits.

This fixes cpv.py and query.py to allow package names of all digits.