Fix broken test suite
The test suite (python setup.py test) was depending on a broken
interface from dev-python/snakeoil. This indicates that tests weren't
run in a while.

We don't actually need snakeoil to run tests. Replaced the test command
with a plain distutils Command.

Moreover, one of test Query tests wasn't in sync with the code. Changed
it to reflect current code.