Remove pylint
As of now, pylint checks return a metric ton of warnings, which tells us
that it hasn't been running lately. pylint is replaced by flake8 and its
continuous run will enventually be ensured by a CI running tox on this
project. The immediate goal, for now, is to have an easy command that
checks as much as possible, but that *passes* so that we don't litter
the project with more code that decrease quality.