equery: Add -F TMPL option to depends module
Like in other modules it is now possible to format the
output of the equery module 'depends' with the commandline
switch '-F' and TMPL.

The man page was adujusted.

A new static method 'print_formated' was created.
In method 'format_depend' is checked if the new option is present.
Depending on the check the PackageFormatter is used to display
the dependencies.

manually apply github PR, some minor editing.
Signed-off-by: Brian Dolbec <dolsen@gentoo.org>