belongs.py: replace uprint in BelongsPrinter with print
Both of these uprint calls are printing Python "str" strings, which do
not require the encoding logic in uprint, since they are already
encoded.

The first method is printing the pkg.cpv explicitly casted into a str,
or printing pkg.cp which is already a str.

For the second method, appending a string literal to a byte string causes a runtime
exception, so it is not possible that it prints anything that is not a
str.

The currently existing callsites for these methods currently pass
instances of gentoolkit.package.Package as the pkg parameter. The
Package class's cp property is inherited from gentoolkit.cpv.CPV,
which returns a str.