depends.py: replace uprint in DependsPrinter with print
The print functions are called by the method format_depend, which
always pass str parameters.

The indent and use_conditional variables are created from string literals, which are always
a str in py3.

The cpv parameter is created by calling str() on a field of a class
instance, which always evaluates to str.

The depatom parameter is created by %s formatting a string literal
with another string created with a call to str(), or an expression
that involves appending a str created by a call to str()

All parameters are always of str type in the current gentoolkit, and
str string do not require the legacy text handling in uprint.