revdep_rebuild: Fix issue where library is a full path
app-editors/neovim-0.10.0 has one of it's libraries referenced by a full
path, which seems to be causing it want to rebuild.

On my system scanelf shows:
   /usr/bin/nvim;nvim;;libluv.so.1,libvterm.so.0,/usr/lib64/lua/5.1/lpeg.so,libmsgpack-c.so.2,libtree-sitter.so.0,libunibilium.so.4,libluajit-5.1.so.2,libm.so.6,libuv.so.1,libc.so.6;ELFCLASS64

In the LibCheck.search() function it was passing
"/usr/lib64/lua/5.1/lpeg.so" to LibCheck.check() which searches by
basename, which of course isn't going to match.

Bug: https://bugs.gentoo.org/932671
Part-of: https://github.com/gentoo/gentoolkit/pull/49
Closes: https://github.com/gentoo/gentoolkit/pull/49
Signed-off-by: Sam James <sam@gentoo.org>